refactor: move Target class to new target.py
