+++ gmatpyplus/command/__init__.py
@@ -4,3 +4,4 @@ from .gmat_command import GmatCommand
 from .propagate import Propagate
 from .solver_branch_command import *
 from .solver_sequence_command import *
+from .target import Target
